MultiLine AST node
