+++ lib/keisan.rb
@@ -17,6 +17,7 @@ require "keisan/ast/boolean"
 require "keisan/ast/parent"
 require "keisan/ast/operator"
 require "keisan/ast/assignment"
+require "keisan/ast/multi_line"
 require "keisan/ast/unary_operator"
 require "keisan/ast/unary_identity"
 require "keisan/ast/unary_plus"
@@ -47,6 +48,7 @@ require "keisan/ast/function"
 require "keisan/ast/list"
 require "keisan/ast/indexing"
 
+require "keisan/ast/line_builder"
 require "keisan/ast/builder"
 require "keisan/ast"
