+++ lib/keisan/ast/multi_line.rb
@@ -1,0 +1,21 @@
+module Keisan
+  module AST
+    class MultiLine < Parent
+      def value(context = nil)
+        context ||= Keisan::Context.new
+        evaluate(context).value(context)
+      end
+
+      def evaluate(context = nil)
+        context ||= Keisan::Context.new
+        @children = children.map {|child| child.evaluate(context)}
+        @children.last
+      end
+
+      def simplify(context = nil)
+        context ||= Keisan::Context.new
+        evaluate(context)
+      end
+    end
+  end
+end
